adjust reduce func
